+++ rust/stackablectl/CHANGELOG.md
@@ -10,9 +10,11 @@ All notable changes to this project will be documented in this file.
 - Add `uninstall` subcommand for `demo`/`stack` commands ([#429]).
 - Add confirmation prompt to `install` subcommand for namespace selection ([#429]).
 - Add `--assume-yes` option for running commands non-interactively ([#429]).
+- Support Helm charts sourced from OCI registries in demo/stack manifests ([#440]).
 
 [#429]: https://github.com/stackabletech/stackable-cockpit/pull/429
 [#438]: https://github.com/stackabletech/stackable-cockpit/pull/438
+[#440]: https://github.com/stackabletech/stackable-cockpit/pull/440
 
 ## [1.4.0] - 2026-03-18